requirements.txt
aioredis==2.0.0
alembic==1.7.3
asyncpg==0.24.0
fakeredis==1.6.1
fastapi==0.68.1
fastapi-utils==0.2.1
httpx==0.19.0
imgkit==1.2.2
jinja2==3.0.1
psycopg2-binary==2.9.1
pydantic[email]
pytest-asyncio==0.15.1
pytz
sqlalchemy[asyncio]==1.4.23
uvicorn[standard]==0.15.0